clean up , added public req table
