+++ src/App.scss
@@ -1264,26 +1264,6 @@
   }
 }
 
-
-.warnmodal {
-  height: 200px;
-  width: 350px;
-  display: inline-block;
-  text-align: center;
-
-  h3 {
-    position: absolute;
-    top: 50%;
-    -ms-transform: translateY(-50%);
-    transform: translateY(-50%);
-    text-align: center;
-    left: 0;
-    right: 0;
-    padding: 10px;
-    font-feature-settings: normal !important;
-  }
-}
-
 .warnmodalledger {
   height: 220px;
   width: 650px;
