+++ src/components/OverviewApp/Notary/index.ts
@@ -2,7 +2,6 @@ import CancelProposalTable from "./CancelProposalTable"
 import LargeRequestTable from "./LargeRequestTable"
 import NotaryTabs from "./NotaryTabs"
 import PublicRequestTable from "./PublicRequestTable"
-import Doremi from "./Doremi"
 import VerifiedClientsTable from "./VerifiedClientsTable"
 
-export { CancelProposalTable, LargeRequestTable, PublicRequestTable, VerifiedClientsTable, NotaryTabs, Doremi }
+export { CancelProposalTable, LargeRequestTable, PublicRequestTable, VerifiedClientsTable, NotaryTabs }
